+++ scripts/setup-local-cluster.sh
@@ -140,23 +140,21 @@ trap cleanup SIGINT SIGTERM ERR EXIT
 
 # $1 = rest port
 # $2 = node port
-# $3 = admin port
-# $4 = healthcheck port
-# $5 = node data directory
-# $6 = node log file
-# $7 = node id file
-# $8 = host to listen on
-# $9 = OPTIONAL: additional args to hoprd
+# $3 = healthcheck port
+# $4 = node data directory
+# $5 = node log file
+# $6 = node id file
+# $7 = host to listen on
+# $8 = OPTIONAL: additional args to hoprd
 function setup_node() {
   local api_port=${1}
   local node_port=${2}
-  local admin_port=${3}
-  local healthcheck_port=${4}
-  local dir=${5}
-  local log=${6}
-  local id=${7}
-  local host=${8}
-  local additional_args=${9:-""}
+  local healthcheck_port=${3}
+  local dir=${4}
+  local log=${5}
+  local id=${6}
+  local host=${7}
+  local additional_args=${8:-""}
 
   log "Run node ${id} on rest port ${api_port} -> ${log}"
 
@@ -166,9 +164,6 @@ function setup_node() {
 
   log "Additional args: \"${additional_args}\""
 
-  # Set NODE_ENV=development to rebuild hopr-admin next files
-  # at runtime. Necessary to start multiple instances of hoprd
-  # in parallel
   env \
     DEBUG="hopr*" \
     NODE_ENV="${node_env}" \
@@ -178,9 +173,6 @@ function setup_node() {
     HOPRD_NETWORK_QUALITY_THRESHOLD="0.3" \
     HOPRD_ON_CHAIN_CONFIRMATIONS=2 \
     ${hoprd_command} \
-      --admin \
-      --adminHost "${host}" \
-      --adminPort ${admin_port} \
       --announce \
       --api-token "${api_token}" \
       --data="${dir}" \
@@ -270,11 +262,11 @@ cp -R \
 # }}}
 
 #  --- Run nodes --- {{{
-setup_node 13301 19091 19501 18081 "${node1_dir}" "${node1_log}" "${node1_id}" "${listen_host}"
-setup_node 13302 19092 19502 18082 "${node2_dir}" "${node2_log}" "${node2_id}" "${listen_host}"
-setup_node 13303 19093 19503 18083 "${node3_dir}" "${node3_log}" "${node3_id}" "${listen_host}"
-setup_node 13304 19094 19504 18084 "${node4_dir}" "${node4_log}" "${node4_id}" "${listen_host}"
-setup_node 13305 19095 19505 18085 "${node5_dir}" "${node5_log}" "${node5_id}" "${listen_host}"
+setup_node 13301 19091 18081 "${node1_dir}" "${node1_log}" "${node1_id}" "${listen_host}"
+setup_node 13302 19092 18082 "${node2_dir}" "${node2_log}" "${node2_id}" "${listen_host}"
+setup_node 13303 19093 18083 "${node3_dir}" "${node3_log}" "${node3_id}" "${listen_host}"
+setup_node 13304 19094 18084 "${node4_dir}" "${node4_log}" "${node4_id}" "${listen_host}"
+setup_node 13305 19095 18085 "${node5_dir}" "${node5_log}" "${node5_id}" "${listen_host}"
 # }}}
 
 log "Waiting for nodes bootstrap"